Move char extensions class to Util namespace
